pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "palebluedot"
version = "0.0.2"
dependencies = [
    "scipy",
    "numpy",
    "Pillow",
    "orbit-predictor",
    "pyproj",
    "rasterio",
]

[project.scripts]
palebluedot = "palebluedot.cli:run"

[tool.hatch.build]
packages = ["src/palebluedot"]
